refactor client network to command patert sending to align with server new slyle, fix network game sync start, fix animation sync, fix unit tests, cleanup

+++ source/components/managers/BonusEffectManager.cpp
@@ -207,5 +207,5 @@ size_t BonusEffectManager::TankNameToId(const std::string& name)
 		return 5;
 	}
 
-	return -1;
+	return static_cast<size_t>(-1);
 }
